refactor: 将uint64类型统一改为int64以增强兼容性
为了增强代码的兼容性和一致性，将所有使用uint64类型的字段统一改为int64类型。这一改动涉及多个服务模块，包括博客、用户、测试用例、题单等。修改后，代码在处理ID和相关字段时更加统一，减少了类型转换的复杂性。

+++ internal/app/service/collection/collection.go
@@ -11,7 +11,7 @@ import (
 )
 
 func isPermission(c collection.Collection, reqUser model.ReqUser) error {
-	if c.UserId != uint64(reqUser.Id) && reqUser.Role < entity.RoleAdmin {
+	if c.UserId != int64(reqUser.Id) && reqUser.Role < entity.RoleAdmin {
 		query := querycontext.CollectionQueryContext{}
 		query.Id.Add(int64(c.Id))
 		_, map_, err := collection.Query.SelectOne(query, collection.QueryUserId())
@@ -19,7 +19,7 @@ func isPermission(c collection.Collection, reqUser model.ReqUser) error {
 		if err != nil {
 			return err
 		}
-		if !slices.Contains(userIds, reqUser.Id) {
+		if !slices.Contains(userIds, int64(reqUser.Id)) {
 			return errors.ErrUnauthorized.WithMessage("没有权限操作该题单")
 		}
 	}
